powerpc/opal-irqchip: Fix endian handling of last_outstanding_events
This is a combination of two upstream commits:
  25642e1459ac ("powerpc/opal-irqchip: Fix double endian conversion")
  036592fbbe75 ("powerpc/opal-irqchip: Fix deadlock introduced by "Fix double endian conversion"")

The OPAL event calls return a mask of events that are active in big
endian format. This is checked when unmasking the events in
opal_handle_events() by comparison with a cached value called
last_outstanding_events. The cached value is stored in big endian format
but the value it's compared with is in CPU endian.

This bug leads to OPAL event delivery being delayed or dropped on some
systems. Symptoms may include a non-functional console.

The bug is fixed by converting the events value to CPU endian as soon as
we get it from OPAL.

Fixes: 9f0fd0499d30 ("powerpc/powernv: Add a virtual irqchip for opal events")
Cc: stable@vger.kernel.org # v4.2+
Reported-by: Douglas L Lehr <dllehr@us.ibm.com>
Signed-off-by: Alistair Popple <alistair@popple.id.au>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>